+++ PKHeXMac/Core/Util/RandUtil.swift
@@ -9,10 +9,10 @@ import Foundation
 
 enum RandUtil {
     static func rand32() -> UInt32 {
-        (UInt.random(in: 0..<(1 << 30)) << 2) | UInt.random(in: 0..<(1 << 2))
+        (UInt32.random(in: 0..<(1 << 30)) << 2) | UInt32.random(in: 0..<(1 << 2))
     }
 
     static func rand64() -> UInt64 {
-        self.rand32 | UInt64(self.rand32() << 32)
+        UInt64(self.rand32()) | UInt64(self.rand32() << 32)
     }
 }
